First attempt at multiplayer syncing
The approach:
The click on the update button makes the current client send an upgrade request to the server. The server handles said request, upgrades the production and notifies all clients of that upgrade.
The initial sync is done by iterating over all productions in all farms, and if they are upgraded, send the upgrade event along the syncing connection. This should theoretically sync the mod data while the game displays the synchronizing player message.