Make the mod compatible with old mod saves
The mod previously used the x and y coordinates, which was changed to x and z to properly locate buildings, but old saves don't have the z coordinate, so it gets loaded as nil and results in the savegame not loading.

This change attempts to still load these entries with the x and y coordinate, and the mod save will be fixed when the game is next saved.